requirements.txt
black
dnspython
google-auth
google-oauth
python-dotenv
python-json-logger
requests
pyyaml
